Fix #33 rename android package
