+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,1 @@
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reactlibrary">
-
-</manifest>
-  
+<manifest package="com.uport.signer" />
